requirements.txt
SQLAlchemy==2.0.4
python-dotenv==0.21.1
fastapi==0.92.0
pydantic==1.10.5
pydantic[email]
alembic==1.9.4
asyncpg==0.27.0
uvicorn==0.20.0
httpx==0.23.3
pytest==7.2.2
pytest_asyncio==0.20.3
bcrypt==4.0.1
PyJWT==2.6.0
python-jose==3.3.0
factory_boy==3.2.1